requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==4.0.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
Django==5.2.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-dataclasses==1.4.0
djangorestframework_simplejwt==5.4.0
djoser==2.3.3
drf-yasg==1.21.8
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.8
marshmallow==3.26.1
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.97.1
orjson==3.11.0
packaging==25.0
pillow==11.1.0
pluggy==1.5.0
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.1
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.0
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stripe==12.4.0
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
